int64

+++ AI3/hasi8.99/common.hpp
@@ -6,26 +6,26 @@
 #include <sstream>
 using namespace std;
 
-using Edge = pair<int, int>;
+using Edge = pair<int64_t, int64_t>;
 
 struct Move {
-    int punter_id;
-    vector<int> route;
+    int64_t punter_id;
+    vector<int64_t> route;
     Move() {}
-    Move(int punter_id, const vector<int>& route) : punter_id(punter_id), route(route) {}
+    Move(int64_t punter_id, const vector<int64_t>& route) : punter_id(punter_id), route(route) {}
 };
 
 using Moves = vector<Move>;
 
 struct Site {
-    int id;
+    int64_t id;
     double x, y;
 };
 
 struct Graph {
-    map<int, Site> sites;
+    map<int64_t, Site> sites;
     vector<Edge> edges;
-    vector<int> mines;
+    vector<int64_t> mines;
     string toString() const {
         stringstream ss;
         // 辺数
@@ -37,7 +37,7 @@ struct Graph {
         // mine数
         ss << mines.size() <<' ';
         // mine
-        for(int m : mines){
+        for(int64_t m : mines){
             ss << m <<' ';
         }
         return ss.str();
@@ -49,21 +49,21 @@ Graph parseGraph(string in){
     stringstream ss(in);
 
     // 辺数
-    int en;
+    int64_t en;
     ss >> en;
     // 辺
-    for(int i = 0; i< en; i++){
+    for(int64_t i = 0; i< en; i++){
         Edge e;
         ss >> e.first >> e.second;
         g.edges.push_back(e);
     }
 
     // mine数
-    int em;
+    int64_t em;
     ss >> em;
     // mine
-    for(int i = 0; i< em; i++){
-        int m;
+    for(int64_t i = 0; i< em; i++){
+        int64_t m;
         ss >> m;
         g.mines.push_back(m);
     }
